+++ app/providers/__init__.py
@@ -7,6 +7,8 @@ from .base import Provider
 from .bnf import BnfProvider
 from .googlebooks import GoogleBooksProvider
 from .isbndb import IsbndbProvider
+from .k10plus import K10plusProvider
+from .loc import LocProvider
 from .openbd import OpenBdProvider
 from .openlibrary import OpenLibraryProvider
 from .sudoc import SudocProvider
@@ -16,6 +18,8 @@ __all__ = [
     "BnfProvider",
     "GoogleBooksProvider",
     "IsbndbProvider",
+    "K10plusProvider",
+    "LocProvider",
     "OpenBdProvider",
     "OpenLibraryProvider",
     "SudocProvider",
@@ -33,6 +37,8 @@ def build_providers(settings: Settings) -> list[Provider]:
         BnfProvider.name: BnfProvider(),
         SudocProvider.name: SudocProvider(),
         OpenLibraryProvider.name: OpenLibraryProvider(),
+        K10plusProvider.name: K10plusProvider(),
+        LocProvider.name: LocProvider(),
         OpenBdProvider.name: OpenBdProvider(),
         GoogleBooksProvider.name: GoogleBooksProvider(settings.google_books_key),
         IsbndbProvider.name: IsbndbProvider(settings.isbndb_key),